requirements.txt
requests
pytest
pyyaml